Make parsing module, move new parser code there
Having things inside the _c_wrappers module causes issues
with curcular imports, and is not good organization.
