+++ src/ert/_c_wrappers/config/__init__.py
@@ -1,5 +1,6 @@
 from .config_content import ConfigContent, ContentItem, ContentNode
-from .config_parser import ConfigParser, ConfigValidationError, ConfigWarning
+from .config_parser import ConfigParser
+from ...parsing import ConfigWarning, ConfigValidationError
 from .config_path_elm import ConfigPathElm
 from .content_type_enum import ContentTypeEnum
 from .schema_item import SchemaItem
@@ -14,6 +15,4 @@ __all__ = [
     "ContentItem",
     "ContentNode",
     "ConfigParser",
-    "ConfigValidationError",
-    "ConfigWarning",
 ]
