+++ src/ert/_c_wrappers/config/config_parser.py
@@ -1,5 +1,4 @@
 import os.path
-from typing import Optional
 
 from cwrap import BaseCClass
 from ecl.util.util import StringHash
@@ -7,24 +6,7 @@ from ecl.util.util import StringHash
 from ert._c_wrappers import ResPrototype
 from ert._c_wrappers.config.config_content import ConfigContent
 from ert._c_wrappers.config.unrecognized_enum import UnrecognizedEnum
-
-
-class ConfigWarning(UserWarning):
-    pass
-
-
-class ConfigValidationError(ValueError):
-    def __init__(self, errors: str, config_file: Optional[str] = None) -> None:
-        self.config_file = config_file
-        self.errors = errors
-        super().__init__(
-            (
-                f"Parsing config file `{self.config_file}` "
-                f"resulted in the errors: {self.errors}"
-            )
-            if self.config_file
-            else f"{self.errors}"
-        )
+from ert.parsing import ConfigValidationError
 
 
 class ConfigParser(BaseCClass):
