+++ src/ert/_c_wrappers/enkf/ensemble_config.py
@@ -14,7 +14,7 @@ from ecl.util.util import StringList
 
 from ert import _clib
 from ert._c_wrappers import ResPrototype
-from ert._c_wrappers.config.config_parser import ConfigValidationError, ConfigWarning
+from ert.parsing import ConfigWarning, ConfigValidationError
 from ert._c_wrappers.config.rangestring import rangestring_to_list
 from ert._c_wrappers.enkf import FieldConfig, GenKwConfig
 from ert._c_wrappers.enkf.config import EnkfConfigNode
