+++ src/ert/_c_wrappers/enkf/ert_config.py
@@ -11,7 +11,7 @@ from typing import Any, ClassVar, Dict, List, Mapping
 import pkg_resources
 
 from ert._c_wrappers.config import ConfigParser
-from ert._c_wrappers.config.config_parser import ConfigValidationError, ConfigWarning
+from ...parsing import ConfigWarning, ConfigValidationError
 from ert._c_wrappers.enkf.analysis_config import AnalysisConfig
 from ert._c_wrappers.enkf.config_keys import ConfigKeys
 from ert._c_wrappers.enkf.ensemble_config import EnsembleConfig
@@ -31,7 +31,7 @@ from ert._clib.config_keywords import init_site_config_parser, init_user_config_
 
 from ._config_content_as_dict import config_content_as_dict
 from ._deprecation_migration_suggester import DeprecationMigrationSuggester
-from .lark_parser import parse as lark_parse
+from ert.parsing.lark_parser import parse as lark_parse
 
 logger = logging.getLogger(__name__)
