+++ src/ert/parsing/config_keywords.py
@@ -5,8 +5,8 @@ from typing import Any, List, Mapping, Optional, Union
 
 from pydantic import BaseModel
 
-from ert._c_wrappers.config import ConfigValidationError
-from ert._c_wrappers.enkf.lark_parser_common import FileContextToken, Instruction
+from ert.parsing import ConfigValidationError
+from ert.parsing.lark_parser_common import FileContextToken, Instruction
 
 # These keys are used as options in KEY:VALUE statements
 BASE_SURFACE_KEY = "BASE_SURFACE"
