+++ tests/test_config_parsing/test_ert_config_parsing.py
@@ -9,7 +9,7 @@ from textwrap import dedent
 import pytest
 from hypothesis import assume, given
 
-from ert._c_wrappers.config.config_parser import ConfigValidationError, ConfigWarning
+from ert.parsing import ConfigWarning, ConfigValidationError
 from ert._c_wrappers.enkf import ErtConfig
 from ert._c_wrappers.enkf.config_keys import ConfigKeys
