+++ tests/test_config_parsing/test_forward_model.py
@@ -9,7 +9,7 @@ from textwrap import dedent
 import pytest
 from hypothesis import given
 
-from ert._c_wrappers.config.config_parser import ConfigValidationError, ConfigWarning
+from ert.parsing import ConfigWarning, ConfigValidationError
 from ert._c_wrappers.enkf import ErtConfig
 
 from .config_dict_generator import config_generators, to_config_file
