+++ tests/unit_tests/c_wrappers/res/enkf/config/test_gen_kw_config.py
@@ -4,7 +4,7 @@ from textwrap import dedent
 
 import pytest
 
-from ert._c_wrappers.config import ConfigValidationError
+from ert.parsing import ConfigValidationError
 from ert._c_wrappers.enkf import EnKFMain, ErtConfig, GenKwConfig
