+++ tests/unit_tests/c_wrappers/res/enkf/test_ensemble_config.py
@@ -5,7 +5,7 @@ import pytest
 from ecl.grid.ecl_grid import EclGrid
 from ecl.summary import EclSum
 
-from ert._c_wrappers.config import ConfigValidationError, ConfigWarning
+from ert.parsing import ConfigWarning, ConfigValidationError
 from ert._c_wrappers.enkf import ConfigKeys, EnsembleConfig, ErtConfig
 from ert._c_wrappers.enkf.enums import ErtImplType, GenDataFileType
