+++ tests/unit_tests/c_wrappers/res/enkf/test_model_config.py
@@ -1,6 +1,6 @@
 import pytest
 
-from ert._c_wrappers.config import ConfigValidationError
+from ert.parsing import ConfigValidationError
 from ert._c_wrappers.enkf import ErtConfig
